(PA-1983) Separate shared agent components & settings
Managing special cases for agent runtimes while sharing a single common
base project provided too many opportunities for error in the order of
inclusion for settings and components. This commit separates the shared
configuration into two separate files - one each for settings and
components - so that settings can be customized in agent projects before
loading the shared components.

Fixes the gem home for rubygem-hocon, which wasn't correctly set on
5.5.x and master due to this problem.