(PA-1996) Make FFI available for most platforms
Some important ruby gems (e.g. bolt, childprocess) require FFI. This
changes the packaging, so that FFI is available on most platforms.

Disabled platforms:

AIX, Solaris and EL5 are not able to build FFI out of the box.

Cisco-wrlinux, Cumulus and EOS are space constrained, and unlikely to
require FFI.