(PA-2055) Build static & shared boost on Windows
On *nix systems, the boost's default "minimal" build strategy (defined
in boost's boostcpp.jam) builds both static and shared libraries, but on
Windows, it builds only static libraries. This sets explicit params for
Windows so that the options match those of *nix systems, and both static
and shared libraries are built everywhere.