(PA-2055) Patch boost to supply -h to the linker on solaris
Solaris machines need to provide the shared object name via the -h flag to the
linker when compiling things. This is so when other things are linked against
the lib just the lib name is listed (and not an entire path). See
https://docs.oracle.com/cd/E19253-01/817-1984/chapter4-97194/index.html for a
better description.

This commit adds a patch to the gcc jamfile in boost.build to force the usage
of the -h flag via SONAME_OPTION.