(maint) Bump minitar version for bolt runtime
Bolt shares a dependency with puppet-runtime on the minitar gem. In order to build pe-bolt-server we were also adding the newer version in the bolt-vanagon project. This resulted in puppet-bolt and pe-bolt-server packages having different minitar versions. This commit allows us to specify the latest minitar version in bolt-runtime without affecting other projects that may use the older gem.