(BOLT-1244) Parity between agent and bolt-runtime
In order to rely on bolt's puppet for applying manifest code on local targets we are pulling in components from puppet-runtime in order to provided parity between bolt and agent runtime. This commit pulls in augeas and some of the other linux specific gems as well as SELinux for EL platforms.