(PA-2191) winruby UTF8 Fallback for no CodePage
This is to deal specifically with setting the Region to Arabic which
sets the codepage to 720. This codepage isn't recognises by Ruby which
causes it to fail during initialisation when setting up LOCALE.

So patch Ruby for Windows Only to Fallback to UTF for THIS ERROR ONLY.
This also applies to any other languages that use a codepage not
recognised by Ruby.