(BOLT-1360) Include a script for checking for gem updates
This commit adds a simplified version of the script https://github.com/puppetlabs/bolt-vanagon/blob/master/generate.rb used to generate a list of gems that could be updated. Instead of modifying component files directly it attempts to parse the version and compares it to the version specified in the gemfile passes as positional argument. The script will alert when a gem from Gemfile.lock should be updated and when it is not mirrored in artifactory.