(maint) Update net-ssh dependencies for pe-installer-runtime
This commit updates the dependencies for pe-installer-runtime, so that
net-ssh is able to properly handle the new default format for private
keys shipped in the version of OpenSSH that comes as the default on el8
platforms. This is (again, unfortunately) for our version of Bolt, which
isn't installed as a package and as such does not come with some of the
handy dependencies that should be present.