(maint) Keep minitar in pdk-runtime at 0.6.1
PDK runtime has to be able to run older versions of Puppet. The mingw
Puppet gems have a runtime dependency of minitar ~> 0.6.1 which is not
satisfied by minitar 0.9.

This somewhat reverts the change in #220, but the default behaviour is
now to include minitar 0.9 (so that this does not affect any other
runtimes) and allowing PDK to opt in to minitar 0.6.1.