(PE_25719) Add augeas libs to the pe-installer-runtime
The puppet runtime in pe-installer was not being built with augeas
libraries. There are classes in pe-modules which use augeas to configure
PE. The migrate_replica bolt plan applies the database profile which
ends up involving a pe_postgresql class that uses augeas on some
platforms, and this was failing.