(PA-2985) Build ruby-selinux for Debian and Ubuntu
Since selinux is also available on Debian and Ubuntu, we can build the
ruby-selinux library for these platforms as well.

To build the selinux library we need the SELinux development
headers (`libselinux1-dev`) and `swig`, which is a build dependency.

We can't build on ubuntu-16.04-ppc64el since SELinux headers are not
included in the pre-compiled pl-build-tools packages. They seem to be
shipped only for EL.