(FACT-1946) Build rubygem-ffi on Solaris
Add support for building the FFI gem on cross-compiled Solaris platforms
(i386/x86_64 to sparc).

This can be achieved by monkey-patching rubygems and using a custom
rbconfig to bypass version checks. This is necessary because in
cross-compile situations we install gems with the system ruby which is
usually an older version.

This commit also adds support for non cross-compiled Solaris which
should work out of the box.