(maint) Ignore symbol not in libselinux 2.9
libselinux(-devel) was upgraded in our EL mirrors (from 2.8 to 2.9),
which made a new function appear: `security_reject_unknown`.

When requiring the `selinux` library with an older version of
libselinux, the following error appears:

```
undefined symbol: security_reject_unknown - ... selinux.so (LoadError)
```

To ensure our Ruby bindings work with older versions of libselinux, we
omit exporting the newer function to the Ruby library.