(PA-3244) Update boost to version 1.73.0
Boost 1.69 contained a bug where there was a possibility for websocketpp to
enter an infinite loop due to the code expecting a boost call to fail when it
wasn't.

https://github.com/chriskohlhoff/asio/issues/261 has more details on the
specific boost call that failed, but there are cases where websocketpp should
be failing when calling async_read because a socket is closed, but the boost
call would not fail.

This commit updates the boost library to latest (1.73.0) since there are fixes
in newer boost versions to remedy the underlying issue.

This commit updates the cross-compiled builds for boost to build the 'b2'
build tool with a system gcc, rather than pl-gcc. The 'b2' build tool requires
a c++11 compatible gcc, so we are forced to use the system g++ to make things
work.

Once b2 has finished building, the actual boost library build is still done
with the cross-compiler pl-gcc toolchain.

Use a newer GCC to compile the b2 boost build tool, and add the
`pthread` flag so the compiler is able to find `std::thread`.

To avoid pulling tons of dependencies when installing GCC, we do it with
yum. To use yum however, a newer openssl is required. We also install
that.

In the future it might be worth it to transition the other RPMs to be
installed with yum. And maybe mirror the repositories. Whatever.

(PA-3244) Fix build on Solaris

Update patches to be compatible with boost 1.73.0
Remove unused patches
Update ldd flags