(PA-3244) Update cross-compiled boost build
This commit updates the cross-compiled builds for boost to build the 'b2'
build tool with a system gcc, rather than pl-gcc. The 'b2' build tool requires
a c++11 compatible gcc, so we are forced to use the system g++ to make things
work.

Once b2 has finished building, the actual boost library build is still done
with the cross-compiler pl-gcc toolchain.