(PA-3244) Fix boost build on AIX
Use a newer GCC to compile the b2 boost build tool, and add the
`pthread` flag so the compiler is able to find `std::thread`.

To avoid pulling tons of dependencies when installing GCC, we do it with
yum. To use yum however, a newer openssl is required. We also install
that.

In the future it might be worth it to transition the other RPMs to be
installed with yum. And maybe mirror the repositories. Whatever.