(PA-3250) remove env check for OPENSSL_FIPS
OPENSSL_FIPS env variable was used inside openssl 1.0.2 application
to throw error if openssl was not compiled in fips mode.

After move to openssl 1.1.0/1.1.1 branch (that is not supporting FIPS mode)
the check was change to dirrectly throw error if OPENSSL_FIPS is set.

After this commit, the check is removed for redhat7 vendored openssl fips