(maint) Add bundle path to bolt-runtime
In order to run a rake task to generate the Bolt cmdlets Powershell
module in bolt-vanagon we need to use bundler to install gems. This adds
the platform-dependent bundle-path setting to bolt-runtime so that we
can easily call bundler from the bolt component to create a ruby
environment and run the rake task.