(PA-3347) Pull some upstream rubygems commits
There are a lot of deprecation warnings coming from gemspecs that show
up everytime puppet runs, related to the rubyforge_project gemspec
setting.

Pull the commits from rubygems/rubygems#3506 which changed the
deprecation warnings to only show on build. This change will most likely
get into Ruby 2.7.2.

This issue only manifests on cross-compiled platforms where we install
the gems with an older version of RubyGems. On regular platforms the
line does not show up in the gemspec.

This patch should be removed once Ruby 2.7.2 is released and we upgrade
to it.