(PA-3433) Patch augeas to allow AD groups in sudoers
Original commit description:

Before this commit, `/etc/sudoers` files containing AD users or groups
could not be parsed with the sudoers lens because it was containing `\\`
in said users/groups name. Running `visudo -c` shows that a sudoers file
containing these is valid so this case was added in the affected regex.

Augeas PR: https://github.com/hercules-team/augeas/pull/696