(PA-3744) Fix siteruby removal on el-7-aarch64
When installing our compiled ruby we still use some of the code from the
local ruby installation on cross-compiled platforms. On el-7-aarch64
ruby 2.0 expects sitedir to point to a valid path. It's not used for
anything so just make it point to vendordir as well.

We might get hit with this again if we backport the sitedir removal on
6.x, as we have more cross-compiled platforms there... but until then
  this workaround should do.