(PE-31890) Update pe-installer-runtime for main branch
Previously when we built pe-installer-runtime it was always with Ruby 2.5.9.
This works for the 2019.8.x LTS branch of PE, but for the main branch that Ruby
version is a bit behind.
This PR creates a main pe-installer-runtime file and renames the old pe-installer-runtime
file to indicate that is for 2019.8.x.