(maint) Fix gcc installation from AIX Toolbox
IBM recently updated their AIX Toolbox gcc packages, including the
absolute path of libgcc_s.a (from lib/gcc/powerpc-ibm-aix6.1.0.0 to
lib/gcc/powerpc-ibm-aix7.1.0.0).

The absolute path is pointed to by the symlink at
/opt/freeware/lib/libgcc_s.a which is handled by the libgcc-8-1 package,
which doesn't update the link from 6.1.0.0 to 7.1.0.0.

This causes any binary that depends on libgcc_s.a (yum, gcc) to fail.

As a quick workaround, manually overwrite the symlink to point to the
correct library.