(PE-32928) Ship PE Bolt Server with pre-release Puppet gem
This updates the PE Bolt server runtime projects to set a pre-release
version of the Puppet 6.x series gem, and additionally updates the
Puppet rubygem component definition to use Artifactory when pulling in
the pre-release version instead of rubygems, since the gem won't be
published there.

This is in order to pull in changes are loading scripts in modules in
Puppet that are required for testing work in Orchestrator around loading
scripts. For now we plan to pin to this gem version since it should have
all of the changes we need, and doesn't need to be updated with new
changes.

Once Puppet 6.25 is availabe these changes can be removed.