(maint) Make boost and yaml-cpp optional
This commit alters the project file for agent-runtime to make boost and
yaml-cpp optional using the `NO_PXP_AGENT` env variable which was
already available and honored in the puppet-agent project itself.

If `NO_PXP_AGENT` is defined, the project will skip boost and yaml-cpp
since facter does not need those anymore.