(PA-4440) Build java cert store whenever java is available
Previously, we used a long regex to determine which agent platforms
we would build a cert store containing external ca roots for. However,
the allowlist regex is rarely updated and is significantly out of date.

The list is now a blocklist of only FIPS builds, because PE FIPS
uses the BouncyCastle FIPS keystores which would require pulling that
library into the puppet-ca-bundle dependencies to build. It is assumed
that FIPS users would not value easily connecting to the larger Internet
anyways, so we have deferred the work for now. FIPS users may still
create allowed external CAs to connect to but they should use the
Puppet[:ssl_trust_store] setting instead.

For non-FIPS agents, to prevent the list becoming out of date this patch
also updates us to conditionally build the keystore when java is present
intead.