(PA-4494) Update agent runtime's augeas to 1.13.0
The location for augeas' tarball has changed. The previous versions are still hosted
where we expect, but going forward we'll need to grab them from the Augeas Github repo.
This adds a bit more complexity in the augeas component, once all projects are on 1.13.0
we should be able to simplfy things.