(PA-4494) Bump Augeas to 1.1.13
This PR bumps Augeas for all platforms except for Solaris. The readline that is bundled
with Solaris is pretty old and Augeas has a function call that is not defined by readline on
Solaris.
So we are choosing not to bump Augeas for Solaris but for every other platform we can bump it.