(PA-4715) Add Ruby 3.2 to puppet-runtime, validated with EL-7 and EL-8 x86_64 agent-runtime
This PR adds ruby 3.2 as a component to puppet runtime. It also copies the agent-runtime-main
into a new project that has ruby 3.2. Until we have a branch of puppet-agent to promote ruby 2.7.6
to we'll need to seperate out this.
There are three changes that are needed to get Ruby 3.2 to build:
  * Package libffi
  * Package libyaml
  * Patch ruby-shadow
Starting with Ruby 3.2 libffi and libyaml are no longer packaged with Ruby, so we need to take care of that.
Ruby-shadow is a gem dependcy that is way out of date and wont build with Ruby 3, there is a simple patch
we apply that allow us to build it.

The scope of this PR is just to build against EL-7 and EL-8, there is further work to do to get all other
supported platforms to build.