(PA-4720) Removes Racc dependency for macOS
Racc was added as a component in this commit:
https://github.com/puppetlabs/puppet-runtime/commit/8f46e488daab4f1be97ce45f65975c6a8d1d8c3c
As a dependency for Nokogiri.

However, Racc is included in the Ruby standard library and not
needed as a separate gem. It was also encountering compilation
errors when building native extensions on Ruby >= 3.1, so this
commit removes the component entirely as it is extraneous and
causing build issues.