(BOLT-1602) Specialize rbinstall_gem_path.patch
The rbinstall_getm_path patch was hardcoding the puppet agent's prefix
into bolt gems when cross compiled.

Because patch files are not ran through a pre-processor we must
specialize the patch contents and choose which patch to apply in the
component config. When choosing which patch to apply we default to the
old agent path unless we are explicitly building a bolt runtime.