(BOLT-1602) Remove `--bindir` from `gem_install` cmd for bolt-runtime
Bolt runtime specifies the bindir flag when installing gems. Puppet does
not and I have a suspicion that providing the flag is masking the fact
bolt-runtime is not setting all the of project settings Puppet is. AND,
that some of those settings bolt-runtime is missing are preventing it
from being cross compiled.

Removing the bindir flag is part of making bolt-runtime and
agent-runtime-main's configuration more similar in hopes that makes
maintaining shared cross compilation targets possible.