(PA-4737) Reset CPP rbconfig
When building ruby 2.7, `RbConfig::CONFIG['CPP']` contains the interpolated
value of `$(CC) -E`. When building for Solaris (10 & 11, x86_64 and SPARC),
the interpolated value is of the form:

    /opt/pl-build-tools/bin/<arch>-pc-solaris2.1x-gcc -E

However, pl-build-tools doesn't exist on the end target. So reset the `CPP`
option to `gcc -E`, similar to what is done for `CC`

This only affects Solaris for now, but should probably be done for all
hosts that are cross-compiled and/or use pl-build-tools.