(PA-4815) Updates fast_gettext gem to 2.2.0
Puppet projects will soon be using Ruby 3.2, which is not compatible
with versions of fast_gettext earlier than 2.1 because of the
deprecated taint checking features.

This commit updates fast_gettext and related dependencies (gettext,
gettext-setup) and r10k, whose gemspec was recently updated to allow
for newer versions of the gettext stack as well.

Newer versions of gettext introduce new dependencies on erubi and
prime, which this commit adds as new components.