(PA-4844) Patch the autogenerated selinux C extension
Calling into selinux from Ruby 3.2 generates a warning:

    selinux.so: warning: undefining the allocator of T_DATA class swig_runtime_data

This is because ruby 3.2 requires C extensions to "either use
rb_define_alloc_func() to overwrite it or rb_undef_alloc_func() to delete
it."[1] The selinux C extension selinuxswig_ruby_wrap.c is autogenerated by swig
from the selinuxswig_i template.

swig 4.1.1 contains the fix for this[2] but we are using an older version on
RHEL7. So patch the autogenerated C extension, which needs to happen after the
code is generated and before it's compiled.

[1] https://github.com/ruby/ruby/blob/6963f8f743b42f9004a0879cd66c550f18987352/doc/extension.rdoc#label-Write+the+C+Code
[2] https://github.com/swig/swig/commit/962f0900018ecb10287b327d261e424fa5ccb755