(PA-4815) Update installer runtime to support gettext versions
The 2019.8.x stream is pinned back to an older version of gettext and its dependencies to support the version of ruby shipped in the puppet 6 stream. Newer PE streams support running with the unpinned version. In order to support the newer configuration the `erubi` and `prime` rubygems are required. This commit adds the gems to installer runtime artifacts to support the new gettext version.