(PA-4769) Anchor ruby version regex
Gem::Ext::ExtConfBuilder.build is responsible for building a gem with native
extensions. It loads RbConfig::CONFIG to determine compiler flags, architecture,
etc. When building the puppet-runtime and cross compiling, we patch
`ExtConfBuilder.build` to prepend our `rbconfig-x.x.x-orig.rb` which contains
configuration for the *target* platform (eg arm64), not the *host* platform (eg
x86_64).

ExtConfBuilder in Ruby 2.0 and 2.1 contain the line:

    cmd = Gem.ruby << "-I" << File.expand_path("../..", __dir__) << File.basename(extension)

while later versions use `Gem.ruby.shellsplit`:

      cmd = Gem.ruby.shellsplit << "-I" << File.expand_path("../..", __dir__) << File.basename(extension)

The shellsplit is necessary because `Gem.ruby` can contain special characters:

    "'a b'".shellsplit
    => ["a b"]

This commit anchors the ruby version regex so that "3.2.0" does not match /2.0/
This way we apply the correct patch. Doing so allows ffi gem to be cross
compiled on macOS 12 x86_64 for macOS 12 ARM.