(PA-4717) Install libssp-0.dll on Windows
Ruby's configure script enables `--stack-protector-strong` on toolchains that
support it. We're currently using mingw-w64-gcc 5.2.0 on Windows which adds a
runtime dependency to libssp-0.dll when building ruby.exe:

     $ objdump -p ruby.exe | grep 'DLL Name'
     DLL Name: x64-msvcrt-ruby320.dll
     DLL Name: KERNEL32.dll
     DLL Name: msvcrt.dll
     DLL Name: libssp-0.dll

So we need to install this dependency as well. Note newer versions of msys2 and
mingw-w64-gcc no longer add the libssp-0 dependency to programs built with stack
protection, so eventually we can remove it, see

https://github.com/msys2/MINGW-packages/issues/13401
https://www.msys2.org/news/#2022-10-10-libssp-is-no-longer-required