(maint) Don't copy facter dll dependencies
We used to copy some gcc dlls into ruby and native facter's bin directory, see
PA-2067. But we no longer define `facter_root` so we were copying the dlls into
cygwin's `/lib` directory on the build host. If you tried to rebuild on the same
host, it would fail because the destination file existed.