(maint) Export CYGWIN and MAKE environment variables
The :cygwin setting is defined in _shared-agent-settings and the :make setting
is defined in the windows platform definitions, so we should export those values
when compiling openssl 3.0.x like we do for other components so that recursive
make consistently runs the expected make.

The :cc and :cxx settings were originally defined in puppet-agent[1] before the
existence of puppet-runtime. They were later removed[2] and their absence
doesn't affect the openssl 3 component, so remove them.

[1] puppetlabs/puppet-agent@13a2a626b8252e99df2f9d485c4e577309127b37
[2] puppetlabs/puppet-agent@bd5874b7cb7e118d3120ac9d233f8a735dba3b09