(PA-5617) Upgrade to openssl 1.1.2 and curl 7.86 on AIX 7.2
Previous attempts to build yaml-cpp failed when trying to run
/opt/freeware/bin/cmake, because it linked to libcurl.a and generated:

    0509-150 Dependent module /opt/freeware/lib/libcurl.a(libcurl.so.4) could not be loaded.
    0509-103 The module has an invalid magic number.

That library is preinstalled on our AIX 7.2 hosts and is version 7.52.1. The
library is also used by yum/python in order to download AIX toolbox packages.

As mentioned in https://github.com/eclipse-openj9/openj9/issues/4245 upgrading
the curl version resolves the issue. And in order to do that we need to boostrap
a newer openssl 1.1.x version.

Installing newer curl also means we can remove the hack that was added when
building curl.