(PE-36344) Disable curl's ntlm if OpenSSL excludes it
OpenSSL 3.0 by default disables some legacy algorithms that are required
for ntlm. We previously kept them enabled for Bolt's WinRM transport but
in b85bf0fda made the inclusion of these legacy algorithms optional so
that the agent and company no longer have to include insecure crypto on
account of Bolt.

However, removing the NTLM algorithms from OpenSSL caused
downstream failures in the compilation of curl, which this patch
addresses by disabling curl's support for NTLM.