(PA-4867) Build ruby on Solaris 11 SPARC
I couldn't find a way to to make solaris-11-sparc platform definition take into
account the project, because of the way vanagon creates the Platform and then
later the Project. So create a new solaris-113-sparc platform definition for
native Solaris compiles.

Since the name of the file contains 113, vanagon assumes that's the
`os_version`. Setting the os_version from within the platform definition doesn't
seem to work, so I had to specialy case some logic to check for the
platform.name.

Although we're native compiling on SPARC, we have to explicitly set
`--with-baseruby=no` so that ruby's configure script does not attempt to use the
ancient ruby already installed. Instead build miniruby and have the build use
that. We should probably be setting `--with-baseruby=no` always, except in cases
where we're cross compiling to avoid this same problem on other platforms.

Similarly, ruby's configure script will attempt to enable `dtrace` because the
executable is present. Explicitly disable it.