(maint) Add DES back to OpenSSL 3
We ship pgcrypto in PE and it uses the agent runtime's openssl. pgcrypto
hardcodes DES support into it and will fail to start if the DES
algorithm isn't available.

In the future either postgresql should ship its own openssl or we should
patch pgcrypto to not hardcode DES into itself.