(PA-5701) Generate tarballs with solaris-11-sparc in the name
Since the puppet-runtime repo contains multiple projects and we need to
build solaris 11 sparc differently for 7.x vs main, I was using
solaris-113-sparc to fake out vanagon. However, the generated tarballs
still had 113 in the name, which meant pxp-agent-vanagon couldn't find
it when building solaris-11-sparc.

This leverages new functionality in vanagon that allows the platform
name to be overridden based on the string passed to the `platform` DSL
method instead of relying on the platform's filename.

It also undoes some hacks that were in place to differentiate between
solaris 11 and 113.