(PE-36907) Ensure the version of concurrent-ruby matches puppet-agent 7.y
The puppet 8 puppet packages contain the 1.1.10 version of concurrent ruby. Bolt 3.x is still puppet 7 based. In the case where an apply is used over the local transport where the bundled-ruby options is not set it is important that the concurrent ruby versions match. This commit ensures bolt-packages have the same version as puppet 7 packages. As we figure out the bolt 4 / puppet 8 story we can update accordingly.