Drop openjdk build dependency
Native facter 3.x relied on JNI to bridge between Java and native code so that
facter could be required in JRuby/puppetserver[1]. We also relied on javac and
javah when building facter.jar. As a result, the puppet-agent project required
openjdk in order to build native facter[2]. Note this was before the existence
of the puppet-runtime project.

In 2015, we added the ca-cert component to puppet-agent[3]. It was later
renamed to puppet-ca-bundle.

In 2017, we extracted ruby, curl, puppet-ca-bundle, etc into a separate
puppet-runtime project, and the openjdk dependency was copied[4].

In 2020, we dropped support for facter 3 due to puppet 6 going EOL.

In 2022, we no longer run keytool when building the runtime[5]. Instead we
update the puppet-cacerts keystore whenever the puppet-ca-bundle component is
updated. When building the runtime, we just copy over the certs.pem and java
keystore files.

Now that facter 3 is EOL and we no longer need keytool when building
puppet-runtime, we can drop the openjdk dependency.

Closes #751

[1] puppetlabs/facter@b1c916caf6d35fa4da236c3ba7a1ef5ab989969c
[2] puppetlabs/puppet-agent@0da4f8b14addae5926c0186027ae2af5efc20970
[3] puppetlabs/puppet-agent@ee82f3ea0504e875e415825382d3b94a3c1d2f16
[4] puppetlabs/puppet-runtime@a9566bf2881ffc4d832b1c313c7c9eb4d395b440
[5] puppetlabs/puppet-ca-bundle@1b60dd21e1a1d84dda92f65305a60bbc609a97a7