Build pdk runtime components against openssl 3
The order that components are listed in pdk-components matters, so build openssl
3 first followed by other components that rely on it (curl, git, etc). Then
build openssl 1.1.1 before building the extra ruby 2.7.8.

Note the openssl 1.1.1 headers still overwrite the 3.0.x, which is bad and needs
to be resolved somehow as it may confuse any rubygem with native extension that
we install in ruby 2.7.