Refactor how libssp-0.dll is installed into primary ruby's bin
We compile ruby 3.2 with strong stack protection. This causes ruby to be linked
against libssp-0, which must be redistributed with the runtime:

    $ objdump -p ruby.exe | grep libssp
            DLL Name: libssp-0.dll

Windows doesn't have an RPATH equivalent, so the dll must be copied into the
same directory as ruby.

This was partially addressed in commmit ba4c95513, however, the _base-ruby
component is already responsible for installing Windows specific library
dependencies when the main bin dir is different than ruby's bin dir, so
move that logic there.