(PA-5849) Consistently use miniruby when native compiling ruby 3.2
Ruby's build process requires a "base" ruby. We also need an executable ruby
that can install gems into /opt/puppetlabs/puppet/lib.

Previously, native compilation used whatever system ruby happened to be present.
Since each distro ships a different ruby, our build process used everything from
ruby 2.7 to 3.1 and everything in between.

This inconsistency made it more likely for our rubygem components with native
extensions to fail to install.

It also meant bumping to ruby 3.3 in the future could fail because that version
requires base ruby 3.0 or later, but some system rubies are still 2.7.

Additionally, in the past we've made the mistake of using pl-ruby because we
thought we had to.

This commit ensures we use miniruby for all native compiles. Doing so eliminates
all of the issues mentioned.

When cross-compiling, we still have to use either system ruby or pl-ruby.