Apply Ruby patch to enable FIPS with OpenSSL 3
Prior to this commit, attempting to enable fips_mode in Ruby 3.2.3's
OpenSSL gem would cause the error "This version of OpenSSL does not
support FIPS mode (OpenSSL::OpenSSLError)". This was due to OpenSSL
removing the FIPS_mode() function in OpenSSL 3, which Ruby previously
relied on to detect if FIPS was enabled.

Upstream has fixed this incompatibility with OpenSSL 3 in Ruby 3.3. The
contributor has indicated that the fix may be backported to Rubies
3.0-3.2 in the future.

This commit applies Ruby's fix as a patch.

See also:
- ruby/ruby@678d41bc51fe31834eec0b653ba0e47de5420aa0
- https://bugs.ruby-lang.org/issues/20000